Added src\components\slider.py
